requirements.txt
flask
pdfplumber
requests
matplotlib
python-docx
